+++ internal/tutorial/handler.go
@@ -10,8 +10,8 @@ import (
 )
 
 type CompleteRequest struct {
-	Latitude  *float32 `json:"latitude"`
-	Longitude *float32 `json:"longitude"`
+	Latitude  *float64 `json:"latitude"`
+	Longitude *float64 `json:"longitude"`
 }
 
 func CompleteHandler(c *gin.Context) {
